+++ packages/provider-elasticsearch/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.19.3
+* Allow contexture search trees to override ES highlight section
+* Auto include in ES _.source.includes fields specified in the search highlight override
+
 # 0.19.2
 * Refactor slightly the results highlighting